+++ app/connectors_service/connectors/sources/oracle/__init__.py
@@ -1,0 +1,11 @@
+#
+# Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+# or more contributor license agreements. Licensed under the Elastic License 2.0;
+# you may not use this file except in compliance with the Elastic License 2.0.
+#
+
+from .client import OracleClient
+from .datasource import OracleDataSource
+from .queries import OracleQueries
+
+__all__ = ["OracleDataSource", "OracleClient", "OracleQueries"]
